Refactored the AD entity query handlers to remove a bunch of duplicate code.
